Move rendering stuff to separate namespaces
