+++ src/Gui/Rendering/IGame.cs
@@ -1,6 +1,8 @@
 // SPDX-FileCopyrightText: Copyright (c) 2025 Logan Bussell
 // SPDX-License-Identifier: MIT
 
+using NesNes.Gui.Rendering;
+
 internal interface IGame : IUpdatable, IRenderable
 {
 }
