+++ src/Gui/Rendering/IRenderable.cs
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: Copyright (c) 2025 Logan Bussell
+// SPDX-License-Identifier: MIT
+
+namespace NesNes.Gui.Rendering;
+
+internal interface IRenderable
+{
+    void Render(double deltaTimeSeconds);
+}
