+++ src/Gui/Views/ClosableWindow.cs
