+++ src/Gui/Views/ExampleWindow.cs
