+++ src/Gui/Views/IImGuiWindow.cs
